Updated MongoDB + BigQuery Templates with changes suggested
