--- v2/bigquery-to-mongodb/src/main/java/com/google/cloud/teleport/v2/templates/BigQueryToMongoDb.java
@@ -1,104 +1,0 @@
-/*
- * Copyright (C) 2019 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.teleport.v2.templates;
-
-import com.google.cloud.teleport.v2.templates.MongoDbUtils;
-import com.google.cloud.teleport.v2.templates.MongoDbUtils.BigQueryReadOptions;
-import com.google.cloud.teleport.v2.templates.MongoDbUtils.MongoDbOptions;
-import java.nio.charset.StandardCharsets;
-import org.apache.beam.sdk.Pipeline;
-import org.apache.beam.sdk.io.mongodb.MongoDbIO;
-import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO;
-import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO.Write.CreateDisposition;
-import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO.Write.WriteDisposition;
-import org.apache.beam.sdk.options.PipelineOptions;
-import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.options.Validation;
-import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.MapElements;
-import com.google.api.services.bigquery.model.TableRow;
-import com.google.api.services.bigquery.model.TableSchema;
-import org.bson.Document;
-import org.apache.beam.sdk.transforms.SimpleFunction;
-import com.google.api.services.bigquery.model.TableFieldSchema;
-
-
-public class BigQueryToMongoDb {
-  /**
-   * Options supported by {@link PubSubToMongoDB}
-   *
-   * <p>Inherits standard configuration options.
-   */
-
-
-  public interface Options
-      extends PipelineOptions,MongoDbOptions, BigQueryReadOptions  {
-  }
-
-  private static class ParseAsDocumentsFn extends DoFn<String, Document> {
-
-    @ProcessElement
-    public void processElement(ProcessContext context) {
-      context.output(Document.parse(context.element()));
-    }
-  }
-
-
-  public static void main(String[] args) {
-
-    Options options = PipelineOptionsFactory.fromArgs(args).withValidation().as(Options.class);
-    run(options);
-  }
-
-
-  public static boolean run(Options options) {
-    Pipeline pipeline = Pipeline.create(options);
-
-
-    pipeline
-            .apply(
-                BigQueryIO
-                        .readTableRows()
-                        .withoutValidation()
-                        .from(options.getInputTableSpec()))
-            .apply("bigQueryDataset", MapElements.via
-                (new SimpleFunction<TableRow, Document>() {
-                  @Override
-                  public Document apply(TableRow row) {
-                    Document doc = new Document();
-                    row.forEach((key, value) ->
-                    {
-                      if(key != "_id"){
-                        doc.append(key, value);
-                      }
-                    });
-                    return doc;
-                  }
-                })
-            )
-            .apply(
-                    MongoDbIO.write()
-                            .withUri(options.getMongoDbUri())
-                            .withDatabase(options.getDatabase())
-                            .withCollection(options.getCollection())
-            );
-    pipeline.run();
-    return true;
-  }
-
-
-
-}
