+++ v2/googlecloud-to-mongodb/README.md
@@ -1,0 +1,9 @@
+# Google Cloud to MongoDB Dataflow templates
+
+is the top-level module containing a batch
+of templates enabling data integration between Google Cloud products and MongoDB.
+
+## Templates
+* [BigQueryToMongoDb](docs/BigQueryToMongoDb/README.md)
+
+Please refer to the corresponding links above for additional information.
