+++ v2/googlecloud-to-mongodb/src/main/resources/bigquery-to-mongodb-command-spec.json
@@ -1,0 +1,7 @@
+{
+  "mainClass": "com.google.cloud.teleport.v2.mongodb.templates.BigQueryToMongoDb",
+  "classPath": "/template/googlecloud-to-mongodb/*:/template/googlecloud-to-mongodb/libs/*:/template/googlecloud-to-mongodb/classes",
+  "defaultParameterValues": {
+    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\",\"goog-dataflow-provided-template-name\":\"googlecloud-to-mongodb\"}"
+  }
+}
