--- v2/mongodb-to-bigquery-cdc/src/main/resources/mongodb-to-bigquery-cdc-command-spec.json
@@ -1,7 +1,0 @@
-{
-  "mainClass": "com.google.cloud.teleport.v2.templates.MongoDbToBigQueryCdc",
-  "classPath": "/template/mongodb-to-bigquery-cdc/*:/template/mongodb-to-bigquery-cdc/libs/*:/template/mongodb-to-bigquery-cdc/classes",
-  "defaultParameterValues": {
-    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\",\"goog-dataflow-provided-template-name\":\"mongodb-to-bigquery-cdc\"}"
-  }
-}
