--- v2/mongodb-to-bigquery/src/main/resources/mongodb-to-bigquery-command-spec.json
@@ -1,7 +1,0 @@
-{
-  "mainClass": "com.google.cloud.teleport.v2.templates.MongoDbToBigQuery",
-  "classPath": "/template/mongodb-to-bigquery/*:/template/mongodb-to-bigquery/libs/*:/template/mongodb-to-bigquery/classes",
-  "defaultParameterValues": {
-    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\",\"goog-dataflow-provided-template-name\":\"mongodb-to-bigquery\"}"
-  }
-}
