+++ v2/mongodb-to-googlecloud/README.md
@@ -1,0 +1,10 @@
+# Google Cloud to MongoDB Dataflow templates
+
+is the top-level module containing a batch
+of templates enabling data integration between Google Cloud products and MongoDB.
+
+## Templates
+* [MongoDbToBigQuery](docs/MongoDbToBigQuery/README.md)
+* [MongoDbToBigQueryCDC](docs/MongoDbToBigQueryCDC/README.md)
+
+Please refer to the corresponding links above for additional information.
