+++ v2/mongodb-to-googlecloud/src/main/resources/mongodb-to-bigquery-cdc-command-spec.json
@@ -1,0 +1,7 @@
+{
+  "mainClass": "com.google.cloud.teleport.v2.mongodb.templates.MongoDbToBigQueryCdc",
+  "classPath": "/template/mongodb-to-googlecloud/*:/template/mongodb-to-googlecloud/libs/*:/template/mongodb-to-googlecloud/classes",
+  "defaultParameterValues": {
+    "labels": "{\"goog-dataflow-provided-template-type\":\"flex\",\"goog-dataflow-provided-template-name\":\"mongodb-to-googlecloud\"}"
+  }
+}
